% Auto-generated: do not edit by hand
\name{dashSummernote}

\alias{dashSummernote}

\title{DashSummernote component}

\description{

}

\usage{
dashSummernote(id=NULL, height=NULL, toolbar=NULL, value=NULL)
}

\arguments{
\item{id}{Character. }

\item{height}{Numeric. }

\item{toolbar}{Unnamed list. }

\item{value}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

